fix(Headline): change implementation of Headline
It no longer contains both fullname and position but contains a generic
content. Headlines can be nested to create sub headline. This is used to
create small text for position and big text for fullname.

Besides, I also make some minor changes in the scss of page and avatar
to make the UI more beautiful.